requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cryptography==3.4.7
DateTime==4.3
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
idna==2.10
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
zope.interface==5.4.0
gunicorn